test(accordion): update tests
